ims_usrloc_scscf: assignment of length missing for query_buffer in db_link_contact_to_impu()
When writing to query_buffer with the help of the snprintf() function,
the result of the functio is written to variable query_buffer_len
instead of to the query_buffer.len itself. This leads to core dump
in some cases. Replaced "_" by "." in "query_buffer_len ="

(cherry picked from commit 11c179ab232222f1c78f19557032afd49bef7324)